+++ src/contracts/index.ts
@@ -54,3 +54,4 @@ export * from "./EventAnalyzer";
 export * from "./CapitalAllocation";
 export * from "./EventContractObservation";
 export * from "./EventContractShadowLedger";
+export * from "./ResearchIntegrity";